Made changes in the Query and Modified the Model classes and Response classes for the Having all the data related to the posts.

+++ src/main/java/com/trove/response/PostFeedSummary.java
@@ -6,5 +6,5 @@ public interface PostFeedSummary {
     String getPostCaption();
     String getMedia();
     String getPostCreatedTime();
-    Integer getTotalCount();
+    Long getTotalCount();
 }
